requirements.txt
# Flask и расширения
flask==3.1.0
flask-sqlalchemy==3.1.1
flask-migrate==4.1.0
flask-login==0.6.3
flask-wtf==1.2.2
flask-admin==1.6.1
flask-babel==4.0.0
flask-mail==0.9.1
flask-apscheduler==1.13.1
email_validator==2.2.0
python-dotenv==1.1.0
PyJWT==2.8.0
Authlib==1.3.0

# Для WebSocket и интеграции с Figma
websockets==15.0.1
websocket-client==1.7.0

# Утилиты базы данных
alembic==1.15.2
sqlalchemy==2.0.40

# Для двухфакторной аутентификации
pyotp==2.9.0
qrcode==7.4.2
Pillow==10.2.0

# Другие зависимости
Werkzeug==3.1.3
Jinja2==3.1.6
click==8.1.8
wtforms==3.2.1
itsdangerous==2.2.0
Babel==2.17.0
pytz==2025.2
uuid==1.30
requests==2.31.0
cryptography==42.0.2 